validaciones para rutas

+++ src/main/java/com/utp/viacosta/service/impl/RutaServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class RutaServiceImpl implements RutaService {
@@ -26,12 +27,22 @@ public class RutaServiceImpl implements RutaService {
 
     @Override
     public void eliminarRuta(int idRuta) {
-        rutaRepository.deleteById(idRuta);
+        Optional<RutaModel> ruta = rutaRepository.findById(idRuta);
+        if(ruta.isPresent()){
+            rutaRepository.deleteById(idRuta);
+        } else {
+            throw new RuntimeException("No se encontro la ruta con el id: "+idRuta);
+        }
     }
 
     @Override
     public void actualizarRuta(RutaModel ruta) {
-        rutaRepository.save(ruta);
+        Optional<RutaModel> rutaModel = rutaRepository.findById(ruta.getIdRuta());
+        if (rutaModel.isPresent()) {
+            rutaRepository.save(ruta);
+        } else {
+            throw new RuntimeException("No se encontro la ruta con el id: " + ruta.getIdRuta());
+        }
     }
